fix(ci): repair all red CI jobs — RUSTFLAGS clobber, missing artifacts, workspace-wide clippy rot
Root causes across the 11 failing jobs:

- ci.yml global RUSTFLAGS=-D warnings REPLACED (not merged with) the
  per-target rustflags in .cargo/config.toml, dropping relocation-model=pic
  (riscv64), code-model=kernel/static (x86_64) and +bti,+paca,+pacg
  (aarch64) — every link failed with R_RISCV_64/ABS64 relocation errors.
  Removed; warnings stay gated by the clippy jobs' -- -D warnings.
- kernel_fs.img is gitignored, so CI kernels could never embed a VIFS1:
  build.rs now stubs an empty image for compile-only contexts, the Build
  (rv64) job assembles a real bootstrap ramdisk (new
  scripts/build-boot-ramdisk-ci.sh), and the aarch64 boot job builds one
  from its own cells before the kernel step.
- Workspace builds linked cells that need artifacts CI cannot have:
  mlibc-smoke (prebuilt libc.a), doom/tetris-c (source clones),
  lua/tetris-lua (xpack newlib, whose locale objects are non-PIC anyway)
  — excluded from the CI Build matrix.
- littlefs2-sys needs <string.h> that Ubuntu's bare-elf gcc does not ship:
  reuse the vendored third_party/freestanding-include on riscv64
  (lint job, Build job, test-hooks/srv/shell scripts, gen_disk env).
- cargo-geiger cannot analyse -Z build-std targets and its zero-unsafe
  premise is false (Driver Cells need MMIO/DMA unsafe). Replaced with
  scripts/check-cells-unsafe-ratchet.py: a 47-file audited allowlist that
  must never grow (ci.yml + security.yml).
- cargo-deny/audit fail on RUSTSEC-2023-0071 (rsa via embedded-tls, cert
  signature verification only, no fixed release): documented ignore in
  deny.toml + new .cargo/audit.toml; security.yml actions bumped to v2
  for CVSS 4.0 parsing.
- tests/integration pins x86_64-pc-windows-msvc for local dev; CI test
  steps now override with CARGO_BUILD_TARGET=x86_64-unknown-linux-gnu.
- The riscv64 workspace clippy gate had never actually run: fixed the
  full backlog it surfaced (hal-riscv/hal-arm safety docs, lua PIC C
  objects, duplicate _start in test-isolation, net-broker fmt, dead-code
  annotations, ~60 machine-applied lints) plus an aarch64 kernel type
  error (set_kernel_stack u64 vs usize) hidden behind the ramdisk break.

Verified locally: cargo fmt --check, unsafe ratchet, riscv64 release
workspace build with the CI exclusion set, and clippy -D warnings green
on riscv64 workspace + aarch64/x86_64 kernel.

+++ cells/drivers/virtio-gpu/src/cursor.rs
@@ -19,10 +19,10 @@ const SPRITE_LEN: usize = 64 * 64 * 4;
 /// `hot` packs `(hot_x << 16) | hot_y` (hotspot within sprite).
 /// `data_ptr` points to the sprite bytes in the compositor's SAS address space.
 pub fn set_sprite(dev: &mut GpuDevice, data_ptr: usize, xy: u32, hot: u32) {
-    let x = ((xy >> 16) & 0xFFFF) as u32;
-    let y = (xy & 0xFFFF) as u32;
-    let hot_x = ((hot >> 16) & 0xFFFF) as u32;
-    let hot_y = (hot & 0xFFFF) as u32;
+    let x = (xy >> 16) & 0xFFFF;
+    let y = xy & 0xFFFF;
+    let hot_x = (hot >> 16) & 0xFFFF;
+    let hot_y = hot & 0xFFFF;
     // SAFETY: data_ptr is a compositor user-space pointer valid in SAS;
     // SPRITE_LEN is constant and bounds-checked by the VirtIO cursor spec.
     let image = unsafe { core::slice::from_raw_parts(data_ptr as *const u8, SPRITE_LEN) };
@@ -35,7 +35,7 @@ pub fn set_sprite(dev: &mut GpuDevice, data_ptr: usize, xy: u32, hot: u32) {
 ///
 /// `xy` packs `(x << 16) | y`.  Cheap: issues MOVE_CURSOR only (no DMA).
 pub fn move_to(dev: &mut GpuDevice, xy: u32) {
-    let x = ((xy >> 16) & 0xFFFF) as u32;
-    let y = (xy & 0xFFFF) as u32;
+    let x = (xy >> 16) & 0xFFFF;
+    let y = xy & 0xFFFF;
     let _ = dev.gpu.move_cursor(x, y);
 }
